requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# FastAPI
fastapi==0.104.1
uvicorn==0.24.0

# Database
psycopg2-binary==2.9.7
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# OpenAI
openai==1.3.8

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.9.0

# Vector operations
numpy==1.24.3
pgvector==0.2.3

# Environment management
python-dotenv==1.0.0

# Additional utilities
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0